:recycle: Refactor: 빌드 정상화
<footer>
- 관련: #16
