--- app/src/main/java/com/example/saybettereducator/MainActivity.kt
@@ -1,37 +1,0 @@
-package com.example.saybettereducator
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.imePadding
-import androidx.compose.foundation.layout.systemBarsPadding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.rememberNavController
-import com.example.saybettereducator.ui.navigation.AppNavHost
-import com.example.saybettereducator.ui.theme.SaybetterEducatorTheme
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContent {
-            SaybetterEducatorTheme {
-                val navController = rememberNavController()
-                Surface(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .systemBarsPadding()
-                        .imePadding(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    AppNavHost(navController = navController)
-                }
-            }
-        }
-    }
-}
