+++ app/src/main/java/com/example/saybettereducator/config/SayBetterApplication.kt
@@ -1,10 +1,10 @@
-package com.example.saybettereducator
+package com.example.saybettereducator.config
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class App : Application() {
+class SayBetterApplication : Application() {
     override fun onCreate() {
         super.onCreate()
     }
