+++ app/src/main/java/com/example/saybettereducator/ui/view/components/userinfo/UserInfoProfilePopup.kt
@@ -1,4 +1,4 @@
-package com.example.saybettereducator.ui.view.components
+package com.example.saybettereducator.ui.view.components.userinfo
 
 import android.net.Uri
 import androidx.compose.foundation.Image
