+++ app/src/main/java/com/example/saybettereducator/ui/view/components/userinfo/UserInfoTopAppBar.kt
@@ -1,4 +1,4 @@
-package com.example.saybettereducator.ui.view.components
+package com.example.saybettereducator.ui.view.components.userinfo
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
