+++ app/src/main/java/com/example/saybettereducator/ui/view/screen/UserInfoScreen.kt
@@ -31,12 +31,12 @@ import com.example.saybettereducator.ui.intent.UserInfoIntent
 import com.example.saybettereducator.ui.model.UserInfoState
 import com.example.saybettereducator.ui.sideeffect.UserInfoSideEffect
 import com.example.saybettereducator.ui.theme.SaybetterEducatorTheme
-import com.example.saybettereducator.ui.view.components.NameInputBox
-import com.example.saybettereducator.ui.view.components.ProfileImageCard
-import com.example.saybettereducator.ui.view.components.ProfilePopup
-import com.example.saybettereducator.ui.view.components.UserInfoTitleText
-import com.example.saybettereducator.ui.view.components.UserInfoBottomBar
-import com.example.saybettereducator.ui.view.components.UserInfoTopAppBar
+import com.example.saybettereducator.ui.view.components.userinfo.NameInputBox
+import com.example.saybettereducator.ui.view.components.userinfo.ProfileImageCard
+import com.example.saybettereducator.ui.view.components.userinfo.ProfilePopup
+import com.example.saybettereducator.ui.view.components.userinfo.UserInfoTitleText
+import com.example.saybettereducator.ui.view.components.userinfo.UserInfoBottomBar
+import com.example.saybettereducator.ui.view.components.userinfo.UserInfoTopAppBar
 import com.example.saybettereducator.ui.viewmodel.UserInfoViewModel
 import com.example.saybettereducator.utils.image.CameraUtil
 import com.example.saybettereducator.utils.image.GalleryUtil
