--- app/src/main/java/com/example/saybettereducator/utils/MyApplication.kt
@@ -1,9 +1,0 @@
-package com.example.saybettereducator.utils
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class MyApplication : Application() {
-
-}
